requirements.txt
webdriver-manager~=4.0.2
selenium~=4.27.1
requests~=2.32.3
python-dotenv~=1.0.1
Flask~=3.1.0
Flask-Cors~=5.0.0

pycryptodome~=3.21.0
cryptography~=44.0.0
jwt~=1.3.1
APScheduler~=3.11.0
bcrypt~=4.2.1
alembic~=1.14.0
SQLAlchemy~=2.0.36